move packages to pkg
