+++ pkg/client/vault/vault.go
