+++ pkg/lock/local/local.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/nimbolus/terraform-backend/terraform"
+	"github.com/nimbolus/terraform-backend/pkg/terraform"
 )
 
 type LocalLock struct {
